# src/backend/config/swagger-config.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DevEx Despatch API",
    "version": "v1",
    "description": "API for handling Category 2 despatch documents and API key management. <u>Request a key via the Getting Started section below.</u>\n\n<b>⚠️ Note: protected requests (most endpoints) should be accessed using an Api-Key header alongside your request.</b>\n\nIf you have any questions/suggestions/change requests, contact us at devex@platform.tcore.network.\n\nThanks for using :) - DevEx team."
  },
  "servers": [
    {
      "url": "https://devex.cloud.tcore.network",
      "description": "DevEx Despatch API Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Api-Key",
        "description": "API key sent in the Api-Key header."
      }
    },
    "schemas": {
...
      "MissingApiKeyErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Missing API key header. A valid key is required for this endpoint."]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
      "InvalidApiKeyErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Invalid API key. Check it matches the one we issued you."]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
...
      "ForbiddenErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Unauthorised access."]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
...
      "BadRequestOrderCancellationErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Missing required field: advice-id"]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
      "BadRequestFulfilmentCancellationErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Missing required field: fulfilment-cancellation-reason"]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
...
      "NotFoundCancellationErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Failed to find Despatch Advice with id e553cc8e-8b37-4a9b-a0cf-87c34ea70a35."]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
      "NotFoundFulfilmentCancellationErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["No fulfilment cancellation found for id e553cc8e-8b37-4a9b-a0cf-87c34ea70a35."]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      },
...
      "CancelOrderRequest": {
        "type": "object",
        "required": ["advice-id", "order-cancellation-document"],
        "properties": {
          "advice-id": {
            "type": "string",
            "format": "uuid",
            "example": "e553cc8e-8b37-4a9b-a0cf-87c34ea70a35"
          },
          "order-cancellation-document": {
            "type": "string",
            "example": "<OrderCancellation>...</OrderCancellation>"
          }
        }
      },
      "CancelOrderResponse": {
        "type": "object",
        "properties": {
          "order-cancellation": {
            "type": "string",
            "example": "<OrderCancellation>...</OrderCancellation>"
          },
          "order-cancellation-reason": {
            "type": "string",
            "example": "Refund requested"
          },
          "order-cancellation-id": {
            "type": "string",
            "format": "uuid"
          },
          "advice-id": {
            "type": "string",
            "format": "uuid"
          },
          "executed-at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-15T09:00:00.000Z"
          }
        },
        "required": [
          "order-cancellation",
          "order-cancellation-reason",
          "order-cancellation-id",
          "advice-id",
          "executed-at"
        ]
      },
...
      "FulfilmentCancellationRequest": {
        "type": "object",
        "required": ["advice-id", "fulfilment-cancellation-reason"],
        "properties": {
          "advice-id": {
            "type": "string",
            "format": "uuid",
            "example": "e553cc8e-8b37-4a9b-a0cf-87c34ea70a35"
          },
          "fulfilment-cancellation-reason": {
            "type": "string",
            "example": "Unable to fulfil delivery window"
          }
        }
      },
      "FulfilmentCancellationResponse": {
        "type": "object",
        "properties": {
          "fulfilment-cancellation": {
            "type": "string",
            "example": "<FulfilmentCancellation>...</FulfilmentCancellation>"
          },
          "fulfilment-cancellation-reason": {
            "type": "string"
          },
          "fulfilment-cancellation-id": {
            "type": "string",
            "format": "uuid"
          },
          "advice-id": {
            "type": "string",
            "format": "uuid"
          },
          "executed-at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-15T09:00:00.000Z"
          }
        },
        "required": [
          "fulfilment-cancellation",
          "fulfilment-cancellation-reason",
          "fulfilment-cancellation-id",
          "advice-id",
          "executed-at"
        ]
      },
...
      "InternalErrorsArrayResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Internal server error, please try again later."]
          },
          "executed-at": {
            "type": "integer",
            "example": 1772246484
          }
        },
        "required": ["errors", "executed-at"]
      }
    }
  },
  "paths": {
...
    "/api/v1/despatch/cancel/order": {
      "post": {
        "tags": ["Despatch Cancellation"],
        "summary": "Cancel an existing despatch advice",
        "description": "Cancels a despatch advice using its advice-id and an order cancellation XML document.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cancellation stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestOrderCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid issued API key",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingApiKeyErrorsArrayResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidApiKeyErrorsArrayResponse"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "API key does not own the despatch advice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenErrorsArrayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Despatch advice not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occured - try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorsArrayResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Despatch Cancellation"],
        "summary": "Retrieve order cancellation",
        "description": "Retrieves an order cancellation using advice-id or cancellation-id query parameters.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "advice-id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Despatch advice id."
          },
          {
            "name": "cancellation-id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Order cancellation id."
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestOrderCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid issued API key",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingApiKeyErrorsArrayResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidApiKeyErrorsArrayResponse"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "API key does not own the despatch advice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenErrorsArrayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cancellation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occured - try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorsArrayResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/despatch/cancel/fulfilment": {
      "post": {
        "tags": ["Despatch Cancellation"],
        "summary": "Create fulfilment cancellation",
        "description": "Generates and stores a fulfilment cancellation document for a despatch advice.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FulfilmentCancellationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fulfilment cancellation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCancellationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestFulfilmentCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid issued API key",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingApiKeyErrorsArrayResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidApiKeyErrorsArrayResponse"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "API key does not own the despatch advice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenErrorsArrayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Despatch advice not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occured - try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorsArrayResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Despatch Cancellation"],
        "summary": "Retrieve fulfilment cancellation",
        "description": "Retrieves a fulfilment cancellation using advice-id or fulfilment-cancellation-id.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "advice-id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Despatch advice id."
          },
          {
            "name": "fulfilment-cancellation-id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Fulfilment cancellation id."
          }
        ],
        "responses": {
          "200": {
            "description": "Fulfilment cancellation record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCancellationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestFulfilmentCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid issued API key",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingApiKeyErrorsArrayResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidApiKeyErrorsArrayResponse"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "API key does not own the despatch advice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenErrorsArrayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Fulfilment cancellation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundFulfilmentCancellationErrorsArrayResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occured - try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorsArrayResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}